requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
async-lru==1.0.2
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
djangorestframework==3.12.4
entrypoints==0.3
idna==3.2
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.27
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
ordered-set==4.0.2
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
pyaes==1.6.1
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
Pyrogram==1.2.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
selenium==3.141.0
Send2Trash==1.8.0
shadowsocks==2.8.2
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
TgCrypto==1.2.2
tornado==6.1
traitlets==5.1.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
